+++ app/src/main/java/jetsetilly/mandelbrot/RenderCanvas/Buffer.java
@@ -5,5 +5,5 @@ import android.graphics.Bitmap;
 interface Buffer {
     void primeBuffer(Bitmap bitmap);
     void flush(Boolean final_flush);
-    void pushDraw(float cx, float cy, int iteration);
+    void pushDraw(int cx, int cy, int iteration);
 }
